requirements.txt
biopython>=1.81
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
openpyxl>=3.1.0
scikit-learn>=1.0.0 